structured-streaming-progress-tracking/listener/requirements.txt
pyspark==4.0.1
delta-spark==4.0.0
pandas==3.0.0
pyarrow==23.0.0
kafka-python==2.3.0